Using the Tizen application manager to receive information about installed apps and to launch apps. This acts as a template for other app framework adoptions. Using the weston ivi shell to control surfaces. Using the input event manager to dispatch hard key inputs.
This is a bigger update, so I increased the version number from 0.1.1 to 0.2.0.

Some parts are under development.

I created the components "WindowsManager" and "InputEventManager" because they are not defined in AGL right now.
As soon as they are defined, we should switch to them.

Find more information on the official wiki page:
https://wiki.automotivelinux.org/homescreen

Signed-off-by: Bocklage, Jens <Jens_Bocklage@mentor.com>