New concept of dynamically placing surfaces in static defined layout areas.
This is useful if an application wants to render more than one surface.
By default, one surface is placed in one layout render area. With the new homescreen.xml
introspection, apps can place their surfaces as they wish (limited to the boundries
of the assigned layou render area!).

The functionality is not fully implemented, but the shared lib already provides the interface.

Signed-off-by: Bocklage, Jens <Jens_Bocklage@mentor.com>