Fix EGL library name in HomeScreen systemd unit
The LD_PRELOAD of the EGL library in HomeScreen.service's Environment
definition needs to be /usr/lib/libEGL.so.1, as the images do not
contain an undecorated .so symlink.  This had recently been fixed in
the old demo's startup scripts, but seems to have been missed in the
switch to systemd startup.

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ief95db5a6a7e26294a2db919c9efe01616fe7755